+++ src/system/WindowSDL.cpp
@@ -125,13 +125,22 @@ int WindowSDL::initialize() {
         return 0;
     }
 
-    SDL_GameControllerAddMapping("341a0000000000000208000000000000,"
-                                 "USB GAMEPAD 8116,"
-                                 "a:b0,x:b2,start:b7,back:b6,leftstick:b8,rightstick:b9,"
-                                 "leftshoulder:b4,rightshoulder:b5,"
-                                 "dpup:h0.1,dpleft:h0.8,dpdown:h0.4,dpright:h0.2,"
-                                 "leftx:a0,lefty:a1,rightx:a3,righty:a2,"
-                                 "lefttrigger:,b:b1,y:b3,lefttrigger:a4,righttrigger:a4");
+    //! \todo Provide a way for user-defined controller mappings
+    /*
+        SDL_GameControllerAddMapping("341a0000000000000208000000000000,"
+                                     "USB GAMEPAD 8116,"
+                                     "a:b0,x:b2,start:b7,back:b6,leftstick:b8,rightstick:b9,"
+                                     "leftshoulder:b4,rightshoulder:b5,"
+                                     "dpup:h0.1,dpleft:h0.8,dpdown:h0.4,dpright:h0.2,"
+                                     "leftx:a0,lefty:a1,rightx:a3,righty:a2,"
+                                     "lefttrigger:,b:b1,y:b3,lefttrigger:a4,righttrigger:a4");
+        SDL_GameControllerAddMapping("4c050000000000006802000000000000,"
+                                     "PLAYSTATION(R)3 Controller,"
+                                     "a:b14,b:b13,y:b12,x:b15,start:b3,guide:b16,back:b0,"
+                                     "leftstick:b1,rightstick:b2,leftshoulder:b10,"
+                                     "rightshoulder:b11,dpup:b4,dpleft:b7,dpdown:b6,dpright:b5,"
+                                     "leftx:a0,lefty:a1,rightx:a2,righty:a3,lefttrigger:b8,righttrigger:b9");
+    */
 
     for (int i = 0; i < SDL_NumJoysticks(); i++) {
         if (SDL_IsGameController(i)) {